RUST-1780 Bump MSRV to 1.61.0, upgrade `ahash` to 0.8.5 in `MSRV-Cargo.lock`
`ahash` 0.8.3 and older were yanked because of a security vulnerability,
and 0.8.4 (with a bug) and 0.8.5 were released to address that.
`ahash` 0.8.5 requires Rust 1.61+. As a result, the Evergreen task for
`mongo-rust-driver` that checks MSRV compilation started to fail.

There were also a change in the `zerocopy` license,
so `deny.toml` had to be updated.